Test the local plugin
By default, the head of the default branch will be cloned from github.

This change allows to test against local changes to the plugin.
You can set `PLUGIN_URI` to the local directory of the plugin.

In the case of `docker-run-tests.sh` this would just be `.`, as the
plugin is mounted into the container and the workdir is set to the root
dir of this repo.